added the github and form component
