+++ Frontend/package.json
@@ -13,6 +13,7 @@
     "axios": "^1.7.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-router-dom": "^6.24.1",
     "react-toastify": "^10.0.5"
   },
   "devDependencies": {
